refactor: move things
